based off tomas-challenge
the part of code that deals with the white line has been moved out of the main loop so that it only runs until there is no more white line, this saves a bit of processing power (has been tested, no problems). more importantly added a method that detects when the robot is stuck, this method calls another method which should get the robot unstuck. however this does not work atm